Address comments

+++ service/worker/deletenamespace/deleteexecutions/activities.go
@@ -8,6 +8,7 @@ import (
 	"go.temporal.io/api/serviceerror"
 	workflowpb "go.temporal.io/api/workflow/v1"
 	"go.temporal.io/sdk/activity"
+	"go.temporal.io/server/api/adminservice/v1"
 	"go.temporal.io/server/api/historyservice/v1"
 	"go.temporal.io/server/chasm"
 	"go.temporal.io/server/common/dynamicconfig"
@@ -27,7 +28,8 @@ type (
 		visibilityManager manager.VisibilityManager
 		historyClient     historyservice.HistoryServiceClient
 
-		deleteActivityRPS dynamicconfig.TypedSubscribable[int]
+		deleteActivityRPS       dynamicconfig.TypedSubscribable[int]
+		useChasmDeleteExecution dynamicconfig.BoolPropertyFn
 
 		metricsHandler metrics.Handler
 		logger         log.Logger
@@ -66,15 +68,17 @@ func NewActivities(
 	visibilityManager manager.VisibilityManager,
 	historyClient historyservice.HistoryServiceClient,
 	deleteActivityRPS dynamicconfig.TypedSubscribable[int],
+	useChasmDeleteExecution dynamicconfig.BoolPropertyFn,
 	metricsHandler metrics.Handler,
 	logger log.Logger,
 ) *Activities {
 	return &Activities{
-		visibilityManager: visibilityManager,
-		historyClient:     historyClient,
-		deleteActivityRPS: deleteActivityRPS,
-		metricsHandler:    metricsHandler,
-		logger:            logger,
+		visibilityManager:       visibilityManager,
+		historyClient:           historyClient,
+		deleteActivityRPS:       deleteActivityRPS,
+		useChasmDeleteExecution: useChasmDeleteExecution,
+		metricsHandler:          metricsHandler,
+		logger:                  logger,
 	}
 }
 
@@ -194,12 +198,20 @@ func (a *Activities) DeleteExecutionsActivity(ctx context.Context, params Delete
 				NamespaceId:       params.NamespaceID.String(),
 				WorkflowExecution: execution.Execution,
 			})
-		} else {
+		} else if a.useChasmDeleteExecution() {
 			_, err = a.historyClient.DeleteExecution(ctx, &historyservice.DeleteExecutionRequest{
 				NamespaceId: params.NamespaceID.String(),
 				Execution:   execution.Execution,
 				ArchetypeId: archetypeID,
 				Reason:      "Namespace delete",
+			})
+		} else {
+			_, err = a.historyClient.ForceDeleteWorkflowExecution(ctx, &historyservice.ForceDeleteWorkflowExecutionRequest{
+				NamespaceId: params.NamespaceID.String(),
+				ArchetypeId: uint32(archetypeID),
+				Request: &adminservice.DeleteWorkflowExecutionRequest{
+					Execution: execution.Execution,
+				},
 			})
 		}
